MODIFY: model 수정

+++ users/models.py
@@ -5,7 +5,7 @@ from utilities.timestamp import TimeStamp
 class User(TimeStamp): 
     nickname      = models.CharField(max_length=45)
     email         = models.CharField(max_length=50, unique=True)
-    kakao_id      = models.IntegerField()
+    kakao_id      = models.BigIntegerField()
 
     class Meta: 
         db_table = 'users'
